requirements.txt
streamlit
openai
python-dotenv
imapclient
pandas